CI: Add a couple of jobs with interconnect enabled.
Also adds e2e control plane test for interconnect. For now
a single test is added which creates a server pod in one zone
and a client pod in other zone and
  - tests connectivity between the two
  - moves the client pod scheduled node to the server pod schedule node
    zone and tests the connectivity
  - moves back the client pod node to its original zone and tests the
    connectivity again.

This patch also skips e2e control plante tests for egress ip, egress
services and multihoming layer2 and localet networks for multi zone
setup.  Once we add the multi zone interconnect support for these,
we should revert these changes.

+++ test/e2e/egress_services.go
@@ -18,6 +18,7 @@ import (
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/kubernetes/test/e2e/framework"
 	e2enode "k8s.io/kubernetes/test/e2e/framework/node"
+	e2eskipper "k8s.io/kubernetes/test/e2e/framework/skipper"
 	utilnet "k8s.io/utils/net"
 )
 
@@ -36,6 +37,7 @@ var _ = ginkgo.Describe("Egress Services", func() {
 		externalIPv4 string
 		externalIPv6 string
 		nodes        []v1.Node
+		testsSkipped bool
 	)
 
 	f := wrappedTestFramework("egress-services")
@@ -50,6 +52,17 @@ var _ = ginkgo.Describe("Egress Services", func() {
 				"Test requires >= 3 Ready nodes, but there are only %v nodes",
 				len(n.Items))
 		}
+		testsSkipped = true
+		multiZones, err := isMultipleZoneDeployment(f.ClientSet)
+		if err != nil {
+			framework.Failf("Failed to get the node zones : %v", err)
+		}
+		if multiZones {
+			e2eskipper.Skipf(
+				"Egress services are not yet supported with multiple zones deployment",
+			)
+		}
+		testsSkipped = false
 		nodes = n.Items
 		ginkgo.By("Creating an external container to send the traffic to/from")
 		externalIPv4, externalIPv6 = createClusterExternalContainer(externalContainerName, agnhostImage,
@@ -57,7 +70,9 @@ var _ = ginkgo.Describe("Egress Services", func() {
 	})
 
 	ginkgo.AfterEach(func() {
-		deleteClusterExternalContainer(externalContainerName)
+		if !testsSkipped {
+			deleteClusterExternalContainer(externalContainerName)
+		}
 	})
 
 	ginkgotable.DescribeTable("Should validate pods' egress is SNATed to the LB's ingress ip without selectors",
